+++ Tests/Integration/MSFT_xExchOwaVirtualDirectory.Integration.Tests.ps1
@@ -46,6 +46,9 @@ if ($exchangeInstalled)
         return
     }
 
+    # Remove our remote Exchange session so as not to interfere with actual Integration testing
+    Remove-RemoteExchangeSession
+
     Describe 'Test Setting Properties with xExchOwaVirtualDirectory' {
         $testParams = @{
             Identity =  "$($env:COMPUTERNAME)\owa (Default Web Site)"